requirements.txt
alembic==1.11.1 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
automat==22.10.0 ; python_version >= "3.10" and python_version < "4.0"
black==23.3.0 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.6.2 ; python_version >= "3.10" and python_version < "4.0"
brotli==1.0.9 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
cfgv==3.3.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.4 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
constantly==15.1.0 ; python_version >= "3.10" and python_version < "4.0"
cryptography==41.0.1 ; python_version >= "3.10" and python_version < "4.0"
cssselect==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.1.2 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.100.0 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.12.2 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.10" and python_version < "4.0" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64")
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
h2==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
hpack==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.17.3 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.24.1 ; python_version >= "3.10" and python_version < "4.0"
hyperframe==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
hyperlink==21.0.0 ; python_version >= "3.10" and python_version < "4.0"
identify==2.5.24 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
incremental==22.10.0 ; python_version >= "3.10" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
itemadapter==0.8.0 ; python_version >= "3.10" and python_version < "4.0"
itemloaders==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
kaitaistruct==0.10 ; python_version >= "3.10" and python_version < "4.0"
lxml==4.9.2 ; python_version >= "3.10" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.10" and python_version < "4.0"
outcome==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
parsel==1.8.1 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.11.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
pluggy==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
pre-commit==3.3.3 ; python_version >= "3.10" and python_version < "4.0"
protego==0.2.1 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.6 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.0.2 ; python_version >= "3.10" and python_version < "4.0"
pydispatcher==2.0.7 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation == "CPython"
pydivert==2.1.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
pyopenssl==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
pypydispatcher==2.1.2 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation == "PyPy"
pysocks==1.7.1 ; python_version >= "3.10" and python_version < "4.0"
pytest==7.4.0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
queuelib==1.6.2 ; python_version >= "3.10" and python_version < "4.0"
requests-file==1.5.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
scrapy==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
selenium-wire==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
selenium==4.10.0 ; python_version >= "3.10" and python_version < "4.0"
service-identity==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==68.0.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy-cockroachdb==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.18 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
tldextract==3.4.4 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
trio-websocket==0.10.3 ; python_version >= "3.10" and python_version < "4.0"
trio==0.22.1 ; python_version >= "3.10" and python_version < "4.0"
twisted-iocpsupport==1.0.3 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
twisted==22.10.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
urllib3[socks]==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
virtualenv==20.23.1 ; python_version >= "3.10" and python_version < "4.0"
w3lib==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.10" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.10" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
zope-interface==6.0 ; python_version >= "3.10" and python_version < "4.0"
zstandard==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
